requirements.txt
langchain
langsmith
langserve
fastapi
openai
tqdm
python-dotenv
streamlit